Move DependencyImageGenerator to it own file, with its own tests
This change simplifies tests for DependencyGrapGenerator and also adds a
test to check image content by comparing to a golden image.
