+++ src/main/kotlin/com/google/androidstudiopoet/writers/ImageWriter.kt
@@ -1,11 +1,11 @@
 package com.google.androidstudiopoet.writers
 
-import java.awt.image.RenderedImage
+import java.awt.image.BufferedImage
 import java.io.File
 import javax.imageio.ImageIO
 
 class ImageWriter : AbstractWriter {
-    fun writeToFile(content: RenderedImage, path: String) {
+    fun writeToFile(content: BufferedImage, path: String) {
         ImageIO.write(content, "png", File(path))
     }
 }
